requirements.txt
# == CORE ==
fastapi==0.104.0
uvicorn[standard]==0.24.0
pydantic==2.5.0

# == VIRTUALIZATION ==
# libvirt-python==12.0.0  # Requiert libvirt-dev (à installer sur l'hôte KVM)
pyvmomi==8.0.3.0.1      # Pour VMware (simulation)

# == KUBERNETES/OPENSHIFT ==
kubernetes==28.1.0

# == DATABASE ==
sqlalchemy==2.0.0
alembic==1.12.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9

# == UTILITIES ==
python-multipart==0.0.6
pyyaml==6.0.1
httpx==0.25.0
click==8.1.7

# == AUTH ==
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
bcrypt<4

# == OPTIONAL (pour plus tard) ==
# scikit-learn==1.3.0    # Pour l'IA
# pandas==2.1.0          # Pour l'analyse
# paramiko==3.4.0        # Pour SSH avancé
